# back/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /quests:
    get:
      operationId: QuestsController_findAll
      parameters:
        - name: rewardMin
          required: false
          in: query
          description: Prime minimale (incluse)
          schema:
            example: 100
            type: number
        - name: rewardMax
          required: false
          in: query
          description: Prime maximale (incluse)
          schema:
            example: 1000
            type: number
        - name: statusId
          required: false
          in: query
          description: Filtrer par identifiant de statut
          schema:
            example: 2
            type: number
        - name: finalDateBefore
          required: false
          in: query
          description: Date finale avant (ISO 8601)
          schema:
            example: '2025-12-31'
            type: string
        - name: finalDateAfter
          required: false
          in: query
          description: Date finale après (ISO 8601)
          schema:
            example: '2025-01-01'
            type: string
        - name: userId
          required: false
          in: query
          description: Filtrer par identifiant d'utilisateur créateur
          schema:
            example: 1
            type: number
        - name: avgXpMin
          required: false
          in: query
          description: XP moyen minimum des aventuriers (inclus)
          schema:
            example: 10
            type: number
        - name: avgXpMax
          required: false
          in: query
          description: XP moyen maximum des aventuriers (inclus)
          schema:
            example: 100
            type: number
        - name: sortBy
          required: false
          in: query
          description: Champ de tri
          schema:
            example: reward
            type: string
            enum:
              - reward
              - finalDate
              - avgExperience
              - createdAt
        - name: order
          required: false
          in: query
          description: Ordre de tri
          schema:
            example: desc
            type: string
            enum:
              - asc
              - desc
      responses:
        '200':
          description: List of quests (avec filtres et tri)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    name:
                      type: string
                      example: Rescue the Merchant
                    description:
                      type: string
                      example: Escort mission
                    finalDate:
                      type: string
                      format: date-time
                      example: '2025-12-31T23:59:59.000Z'
                    reward:
                      type: number
                      example: 500
                    estimatedDuration:
                      type: number
                      example: 5
                    statusId:
                      type: number
                      example: 1
                    createdAt:
                      type: string
                      format: date-time
                      example: '2025-10-30T12:00:00.000Z'
                    updatedAt:
                      type: string
                      format: date-time
                      example: '2025-10-30T12:34:56.000Z'
      security:
        - bearer: []
      tags:
        - Quests
    post:
      operationId: QuestsController_create
      parameters: []
      requestBody:
        required: true
        description: New quest payload
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                name: Rescue the Merchant
                description: Escort the merchant safely to the city.
                finalDate: '2025-12-31T23:59:59.000Z'
                reward: 500
                estimatedDuration: 5
                adventurerIds:
                  - 1
                  - 2
                equipmentStockIds:
                  - 10
                  - 11
      responses:
        '201':
          description: Quest created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  name: Rescue the Merchant
                  description: Escort the merchant safely to the city.
                  finalDate: '2025-12-31T23:59:59.000Z'
                  reward: 500
                  estimatedDuration: 5
                  statusId: 1
                  createdAt: '2025-10-30T12:00:00.000Z'
                  updatedAt: '2025-10-30T12:00:00.000Z'
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}:
    get:
      operationId: QuestsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Quest by id
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 1
                  title: Rescue the Merchant
                  description: Escort the merchant safely to the city.
                  statusId: 1
                  createdAt: '2025-10-30T12:00:00.000Z'
                  updatedAt: '2025-10-30T12:34:56.000Z'
      security:
        - bearer: []
      tags:
        - Quests
    patch:
      operationId: QuestsController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                name: Rescue the Merchant (Hard)
                description: Hard mode variant.
                reward: 750
      responses:
        '200':
          description: Updated quest
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  name: Rescue the Merchant (Hard)
                  description: Hard mode variant.
                  reward: 750
                  statusId: 1
                  createdAt: '2025-10-30T12:00:00.000Z'
                  updatedAt: '2025-10-30T12:45:00.000Z'
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/status:
    patch:
      operationId: QuestsController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Update quest status
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                statusId: 3
      responses:
        '200':
          description: Quest status updated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  statusId: 3
                  updatedAt: '2025-10-30T12:50:00.000Z'
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/adventurers/attach:
    patch:
      operationId: QuestsController_attachAdventurers
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Attach adventurers to quest
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                ids:
                  - 1
                  - 2
                  - 3
      responses:
        '200':
          description: Adventurers attached
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  adventurerIds:
                    - 1
                    - 2
                    - 3
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/adventurers/detach:
    patch:
      operationId: QuestsController_detachAdventurers
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Detach adventurers from quest
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                ids:
                  - 2
      responses:
        '200':
          description: Adventurers detached
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  adventurerIds:
                    - 1
                    - 3
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/adventurers/set:
    patch:
      operationId: QuestsController_setAdventurers
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Replace all adventurers for quest
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                ids:
                  - 4
                  - 5
      responses:
        '200':
          description: Adventurers set
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  adventurerIds:
                    - 4
                    - 5
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/equipment-stocks/attach:
    patch:
      operationId: QuestsController_attachEquipment
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Attach equipment stocks to quest
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                ids:
                  - 3
                  - 8
      responses:
        '200':
          description: Equipment stocks attached
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  equipmentStockIds:
                    - 3
                    - 8
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/equipment-stocks/detach:
    patch:
      operationId: QuestsController_detachEquipment
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Detach equipment stocks from quest
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                ids:
                  - 3
      responses:
        '200':
          description: Equipment stocks detached
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  equipmentStockIds:
                    - 8
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/equipment-stocks/set:
    patch:
      operationId: QuestsController_setEquipment
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Replace all equipment stocks for quest
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                ids:
                  - 9
                  - 11
      responses:
        '200':
          description: Equipment stocks set
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  equipmentStockIds:
                    - 9
                    - 11
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/validate:
    patch:
      operationId: QuestsController_validate
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: true
        description: Valider la quête et définir l'XP recommandée
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                xp: 150
      responses:
        '200':
          description: Quête validée
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  recommendedXP: 150
                  statusId: 2
                  updatedAt: '2025-10-30T13:15:00.000Z'
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/invalidate:
    patch:
      operationId: QuestsController_invalidate
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      responses:
        '200':
          description: >-
            Quête dévalidée (retour en attente, XP remise à 0,
            aventuriers/équipements/consommables vidés)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  recommendedXP: 0
                  statusId: 1
                  adventurerIds: []
                  equipmentStockIds: []
                  updatedAt: '2025-10-30T13:20:00.000Z'
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/start:
    patch:
      operationId: QuestsController_start
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      responses:
        '200':
          description: Quête démarrée (désélection impossible après démarrage)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  statusId: 3
                  updatedAt: '2025-10-30T13:25:00.000Z'
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/finish:
    patch:
      operationId: QuestsController_finish
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      responses:
        '200':
          description: >-
            Quête terminée, durée de repos calculée automatiquement par
            aventurier (formule SAM)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  statusId: 3
                  adventurers:
                    - id: 1
                      name: Aria
                      experience: 50
                      availableUntil: '2025-12-15'
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/refuse:
    patch:
      operationId: QuestsController_refuse
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: false
        description: Aucun body requis
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: Quête refusée
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  statusId: 4
                  updatedAt: '2025-10-30T13:30:00.000Z'
      security:
        - bearer: []
      tags:
        - Quests
  /quests/{id}/abandon:
    patch:
      operationId: QuestsController_abandon
      parameters:
        - name: id
          required: true
          in: path
          description: Quest ID
          schema:
            example: 42
            type: number
      requestBody:
        required: false
        description: Aucun body requis
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: >-
            Quête abandonnée (peut être abandonnée uniquement si elle n'est pas
            validée ou commencée, typiquement en statut 'en attente')
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                example:
                  id: 42
                  statusId: 5
                  updatedAt: '2025-10-30T13:35:00.000Z'
      security:
        - bearer: []
      tags:
        - Quests
...
info:
  title: API "Les kips AH"
  description: |-
    This is the API information regarding the project "Les capuches d'Opale".

    C'est un outil de gestion d'aventurier pour une guilde.
  version: 1.0.0
  contact: {}
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http